requirements.txt
django==4.0.2
pillow
django_jalali
mysqlclient
unidecode